+++ crates/iceberg/src/lib.rs
@@ -46,3 +46,4 @@ pub mod spec;
 
 pub mod transaction;
 pub mod transform;
+mod scan;
